Processing overlay requests in overlay protocol
